\name{mydata}
\alias{mydata}
\docType{data}
\title{
mydata
}
\description{
A hundred mobility sequences representing 100 unique individuals.  These data are derived from the British Household Panel Study (BHPS)dataset.
}
\usage{data("mydata")}
\format{
  A data frame with 100 observations on the following 14 variables.  
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{X1996}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X1997}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X1998}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X1999}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2000}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2001}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2002}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2003}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2004}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2005}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2006}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2007}}{a factor with levels \code{mover within gb} \code{non-mover}}
    \item{\code{X2008}}{a factor with levels \code{mover within gb} \code{non-mover}}
  }
}


\examples{
data(mydata)
## maybe str(mydata) ; plot(mydata) ...
}
\keyword{datasets}
