\name{myHouseID}
\alias{myHouseID}
\docType{data}
\title{
my house ID's
}
\description{
House ID's for 100 individuals over 13 BHPS census waves.  Each individual is labelled with a personal ID in the first column of the dataset.
}
\usage{data("myHouseID")}
\format{
  A data frame with 100 observations on the following 14 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{X1996}}{a numeric vector}
    \item{\code{X1997}}{a numeric vector}
    \item{\code{X1998}}{a numeric vector}
    \item{\code{X1999}}{a numeric vector}
    \item{\code{X2000}}{a numeric vector}
    \item{\code{X2001}}{a numeric vector}
    \item{\code{X2002}}{a numeric vector}
    \item{\code{X2003}}{a numeric vector}
    \item{\code{X2004}}{a numeric vector}
    \item{\code{X2005}}{a numeric vector}
    \item{\code{X2006}}{a numeric vector}
    \item{\code{X2007}}{a numeric vector}
    \item{\code{X2008}}{a numeric vector}
  }
}
\details{
The individuals represented in the dataset "mydata" are also represented in this dataset ("myHouseID").
}

\examples{
data(myHouseID)
## maybe str(myHouseID) ; plot(myHouseID) ...
}
\keyword{datasets}
